added a fake admin account to the users, created auth.py which authenticates users, deleted the main.py file in the models folder, changed movie route to check for an admin when deleting putting and posing a movie, change the reviewtest file to use the correct status_code variable, can now login on swagger as an admin and unlock deleting, putting and posting but not fully working yet

+++ full-project/backend/app/routers/movieRoute.py
@@ -1,6 +1,8 @@
 
+import json
 from typing import List, Optional
-from fastapi import APIRouter, HTTPException, status
+from app.auth import getCurrentUser
+from fastapi import APIRouter, HTTPException, status, Depends
 from ..schemas.movie import Movie, MovieCreate, MovieUpdate
 from ..services.movieService import (
     listMovies,
@@ -25,20 +27,44 @@ def searchMovies(q: Optional[str] = None, query: Optional[str] = None):
 def getMovies():
     return listMovies()
 
-@router.post("", response_model=Movie, status_code=status.HTTP_201_CREATED)
-def postMovie(payload: MovieCreate):
-    return createMovie(payload)
-
 @router.get("/{movieId}", response_model=Movie)
 def getMovie(movieId: str):
     return getMovieById(movieId)
 
+# -------------- #
+#   ADMIN ONLY   #
+# -------------- #
+
+@router.post("", response_model=Movie, status_code=status.HTTP_201_CREATED)
+def postMovie(payload: MovieCreate, currentUser: dict = Depends(getCurrentUser)):
+    if currentUser["role"] != "admin":
+        raise HTTPException(
+            status_code = status.HTTP_403_FORBIDDEN,
+            detail = "only admins can add movies"
+        )
+    return createMovie(payload)
+
 @router.put("/{movieId}", response_model=Movie)
-def putMovie(movieId: str, payload: MovieUpdate):
+def putMovie(
+    movieId: str, 
+    payload: MovieUpdate,
+    currentUser: dict = Depends(getCurrentUser)): # gets the current user in order to check their role
+    if currentUser["role"] != "admin": # if the current user is not an admin then throw exception
+        raise HTTPException(
+            status_code = status.HTTP_403_FORBIDDEN,
+            detail = "only admins can add movies"
+        )
     return updateMovie(movieId, payload)
 
 @router.delete("/{movieId}", status_code=status.HTTP_204_NO_CONTENT)
-def removeMovie(movieId: str):
+def removeMovie(
+    movieId: str,
+    currentUser: dict = Depends(getCurrentUser)):
+    if currentUser["role"] != "admin": 
+        raise HTTPException(
+            status_code = status.HTTP_403_FORBIDDEN,
+            detail = "only admins can add movies"
+        )
     deleteMovie(movieId)
     return None
